# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application API endpoints protected by Better Auth session cookies. Health is liveness only.",
        "title": "AI Interview Practice API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/jds": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "List current user's JDs",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "maximum": 2147483647,
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Page-domain_ListItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "Save a reviewed JD",
                "parameters": [
                    {
                        "description": "JD request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateJDRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.JDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/jds/analyze": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "Analyze a raw JD without saving",
                "parameters": [
                    {
                        "description": "JD request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AnalyzeJDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.StructuredJD"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/jds/{id}": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "Get an owned JD",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JD UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.JDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "Update reviewed fields; original raw text is immutable",
                "parameters": [
                    {
                        "description": "JD request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateJDRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "JD UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.JDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jds"
                ],
                "summary": "Delete an owned JD",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JD UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperror.Code": {
            "type": "string",
            "enum": [
                "JD_IN_USE",
                "JD_NOT_FOUND",
                "INVALID_CREDENTIALS",
                "USER_NOT_FOUND",
                "USER_INACTIVE",
                "INVALID_TOKEN",
                "AUTH_UNAVAILABLE",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "INTERNAL_ERROR",
                "INVALID_JD_INPUT",
                "JD_TOO_SHORT",
                "JD_TOO_LONG",
                "EXTRACTION_FAILED",
                "INVALID_EXTRACTION_OUTPUT"
            ],
            "x-enum-varnames": [
                "CodeJDInUse",
                "CodeJDNotFound",
                "CodeInvalidCredentials",
                "CodeAccountNotFound",
                "CodeAccountLocked",
                "CodeInvalidToken",
                "CodeAuthUnavailable",
                "CodeForbidden",
                "CodeValidation",
                "CodeInternal",
                "CodeInvalidJDInput",
                "CodeJDTooShort",
                "CodeJDTooLong",
                "CodeExtractionFailed",
                "CodeInvalidExtractionOutput"
            ]
        },
...
        "domain.ExtractedSkill": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/domain.SkillCategory"
                },
                "name": {
                    "type": "string"
                },
                "requirement": {
                    "description": "Empty means the JD does not state whether the skill is required or preferred.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Requirement"
                        }
                    ]
                }
            }
        },
        "domain.ListItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "seniorityLevel": {
                    "$ref": "#/definitions/domain.Seniority"
                },
                "status": {
                    "$ref": "#/definitions/domain.Status"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.Requirement": {
            "type": "string",
            "enum": [
                "required",
                "preferred"
            ],
            "x-enum-varnames": [
                "Required",
                "Preferred"
            ]
        },
        "domain.Seniority": {
            "type": "string",
            "enum": [
                "intern",
                "junior",
                "mid",
                "senior",
                "lead"
            ],
            "x-enum-varnames": [
                "Intern",
                "Junior",
                "Mid",
                "Senior",
                "Lead"
            ]
        },
        "domain.SkillCategory": {
            "type": "string",
            "enum": [
                "programming_language",
                "framework",
                "database",
                "tool",
                "technology",
                "other"
            ],
            "x-enum-varnames": [
                "ProgrammingLanguage",
                "Framework",
                "Database",
                "Tool",
                "Technology",
                "Other"
            ]
        },
        "domain.Status": {
            "type": "string",
            "enum": [
                "uploaded",
                "parsing",
                "parsed",
                "customized",
                "locked",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusUploaded",
                "StatusParsing",
                "StatusParsed",
                "StatusCustomized",
                "StatusLocked",
                "StatusFailed"
            ]
        },
        "domain.StructuredJD": {
            "type": "object",
            "properties": {
                "domainKnowledge": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "seniorityLevel": {
                    "$ref": "#/definitions/domain.Seniority"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ExtractedSkill"
                    }
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.AnalyzeJDRequest": {
            "type": "object",
            "required": [
                "rawText"
            ],
            "properties": {
                "rawText": {
                    "type": "string"
                }
            }
        },
        "handler.CreateJDRequest": {
            "type": "object",
            "required": [
                "rawText",
                "structuredJD"
            ],
            "properties": {
                "rawText": {
                    "type": "string"
                },
                "structuredJD": {
                    "$ref": "#/definitions/domain.StructuredJD"
                }
            }
        },
        "handler.JDResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rawText": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.Status"
                },
                "structuredJD": {
                    "$ref": "#/definitions/domain.StructuredJD"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateJDRequest": {
            "type": "object",
            "required": [
                "structuredJD"
            ],
            "properties": {
                "structuredJD": {
                    "$ref": "#/definitions/domain.StructuredJD"
                }
            }
        },
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.Code"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Page-domain_ListItem": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ListItem"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                }
            }
        },
        "response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Better Auth HttpOnly session cookie. Production uses __Secure-rolecue-authentication.session_token; browser sends cookies automatically.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}